pyproject.toml
[project]
name = "chatbot-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.22.0",
    "openai>=1.68.2",
    "openai-agents>=0.0.6",
    "python-dotenv>=1.0.1",
    "tinydb>=4.8.2",
]
